cleaned up the path handling in the compiler and added a bunch of debugging messages. the compiler can now compile the demo game on Linux.  I tested the rom in an emulator and it seems to work just fine.  now all I have to do is clean up the compile warnings and start cleaning up all of the windows cruft (e.g. FASTCALL).

+++ src/typedefs.h
@@ -12,6 +12,7 @@
 #define _typedefs_h_
 /*********************************************************************/
 
+/* kill FASTCALL...nobody uses this anymore */
 #define FASTCALL
 
 /* data types */
@@ -21,7 +22,7 @@ typedef unsigned short		U16;
 typedef signed short		S16;
 typedef unsigned long		U32;
 typedef signed long			S32;
-typedef unsigned short		BOOL;
+typedef int		            BOOL;
 
 #define TRUE 				1
 #define FALSE 				0
